services/model-config-controller/requirements.txt
kubernetes==29.0.0
psycopg2-binary==2.9.9
